refactor: split into components, get rid of App2
This commit in the initial phase of the app's frontend refactoring
journey. First, the app is now a SPA, with App2.js being completely
removed. FHTWaysApp currently handles showing either the form fields or
the path description based on the component's state.

Furthermore, the entire app's css was refactored. Currently, TailwindCSS
is utilized. The jsx was split into components. This is not the final
version yet, rather initial testing. Next steps include refactoring the
entire app's logic and further refining styling.