refactor: improved input fields and navbar
Input fields now utilize the the MaterialUI Autocomplete & TextField
components. It allows the input field to act both as a regular input
element and as a selector: when a user starts typing in the value, a
list of suggested rooms appears and adjusts itself as the user keeps
typing. However, the user is not bound to picking a value from the list
- they can simply leave their input as is and press submit. This way,
the user both sees the list of allowed nodes and is freed from an
additional action of pressing enter or selecting a particular
input element with a mouth.

Navbar was introduced to include future navigation element (**How it
works** and **Contact us** pages) as well as accessibility buttons.