feature: frontend navbar pages & backend database
On the frontend, the navbar was polished and extended to include two
pages: HowItWorksPage and ContactUsPage. This way, the main FHTWaysPage
responsble for the functionality of the app is not cluttered with
unnecessary data that is delegated to specific pages.

On the backend, a Postgres database was configured and connected with
the Django backend. The db was populated with node and edge data from
the csv files. When the application starts, the Graph is initialized
with the data from the db. This way, the Graph keeps running and waiting
for incoming requests from the frontend, instead of being recreated on
each API call.